backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.0.1
